extract the `Kind*` components to their own files
