+++ client/src/app/pages/identities/components/identity-form/identity-form.tsx
@@ -1,17 +1,14 @@
-import React, { useState } from "react";
-import { Trans, useTranslation } from "react-i18next";
+import React from "react";
+import { useTranslation } from "react-i18next";
 import * as yup from "yup";
 import { AxiosError } from "axios";
 import {
   ActionGroup,
-  Alert,
   Button,
   ButtonVariant,
-  FileUpload,
   Form,
-  Switch,
 } from "@patternfly/react-core";
-import { FormProvider, useForm, useFormContext } from "react-hook-form";
+import { FormProvider, useForm } from "react-hook-form";
 import { yupResolver } from "@hookform/resolvers/yup";
 
 import { OptionWithValue, SimpleSelect } from "@app/components/SimpleSelect";
@@ -23,17 +20,22 @@ import {
   useFetchIdentities,
   useUpdateIdentityMutation,
 } from "@app/queries/identities";
-import KeyDisplayToggle from "@app/components/KeyDisplayToggle";
 import {
   HookFormPFGroupController,
   HookFormPFTextInput,
 } from "@app/components/HookFormPFFields";
 import { NotificationsContext } from "@app/components/NotificationsContext";
+
 import { KIND_OPTIONS } from "../../utils";
+import { KindSourceForm } from "./kind-source-form";
+import { KindAssetForm } from "./kind-asset-form";
+import { KindMavenSettingsFileForm } from "./kind-maven-settings-file-form";
+import { KindSimpleUsernamePasswordForm } from "./kind-simple-username-password-form";
+import { KindBearerTokenForm } from "./kind-bearer-token-form";
 
 import "./identity-form.css";
 
-type UserCredentials = "userpass" | "source";
+export type UserCredentials = "userpass" | "source";
 
 interface IdentityFormValues {
   name: string;
@@ -458,492 +460,5 @@ export const IdentityForm: React.FC<IdentityFormProps> = ({
         </ActionGroup>
       </Form>
     </FormProvider>
-  );
-};
-
-const USER_CREDENTIALS_OPTIONS: OptionWithValue<UserCredentials>[] = [
-  {
-    value: "userpass",
-    toString: () => `Username/Password`,
-  },
-  {
-    value: "source",
-    toString: () => `Source Private Key/Passphrase`,
-  },
-];
-
-const KindSourceForm: React.FC<{
-  identity?: Identity;
-  defaultIdentities?: Record<IdentityKind, Identity | undefined>;
-}> = ({ identity, defaultIdentities }) => {
-  const { t } = useTranslation();
-  const { control, getValues, setValue, resetField } = useFormContext();
-  const values = getValues();
-
-  const [isKeyFileRejected, setIsKeyFileRejected] = useState(false);
-
-  const [isPasswordHidden, setIsPasswordHidden] = useState(true);
-  const toggleHidePassword = (e: React.FormEvent<HTMLElement>) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsPasswordHidden(!isPasswordHidden);
-  };
-
-  const isPasswordEncrypted = identity?.password === values.password;
-  const isKeyEncrypted = identity?.key === values.key;
-  const kindDefault = defaultIdentities?.[values.kind as IdentityKind];
-  const isReplacingDefault =
-    values.default &&
-    kindDefault &&
-    (!identity || kindDefault.id !== identity.id);
-
-  return (
-    <>
-      <HookFormPFGroupController
-        control={control}
-        name="userCredentials"
-        label="User credentials"
-        isRequired
-        fieldId="user-credentials-select"
-        renderInput={({ field: { value, name, onChange } }) => (
-          <SimpleSelect
-            id="user-credentials-select"
-            toggleId="user-credentials-select-toggle"
-            toggleAriaLabel="User credentials select dropdown toggle"
-            aria-label={name}
-            value={
-              value ? toOptionLike(value, USER_CREDENTIALS_OPTIONS) : undefined
-            }
-            options={USER_CREDENTIALS_OPTIONS}
-            onChange={(selection) => {
-              const selectionValue =
-                selection as OptionWithValue<UserCredentials>;
-              onChange(selectionValue.value);
-              // So we don't retain the values from the wrong type of credential
-              resetField("default");
-              resetField("user");
-              resetField("key");
-              resetField("password");
-            }}
-          />
-        )}
-      />
-      <HookFormPFGroupController
-        control={control}
-        name="default"
-        fieldId="default"
-        label="Default credential?"
-        renderInput={({ field: { onChange, value, name, ref } }) => (
-          <>
-            <Switch
-              id="default"
-              name={name}
-              label={t("credentials.default.sourceSwitchLabel")}
-              isChecked={value}
-              onChange={(_, checked) => onChange(checked)}
-              ref={ref}
-            />
-            {isReplacingDefault && (
-              <Alert
-                isInline
-                className="alert-replacing-default"
-                variant="warning"
-                title={t("credentials.default.changeTitle")}
-              >
-                <Trans
-                  i18nKey="credentials.default.sourceChangeWarning"
-                  values={{
-                    name: kindDefault.name,
-                  }}
-                />
-              </Alert>
-            )}
-          </>
-        )}
-      />
-
-      {values?.userCredentials === "userpass" && (
-        <KindSimpleUsernamePasswordForm
-          identity={identity}
-          usernameLabel="Username"
-          passwordLabel="Password"
-        />
-      )}
-
-      {values?.userCredentials === "source" && (
-        <>
-          <HookFormPFGroupController
-            control={control}
-            name="key"
-            fieldId="key"
-            label="Upload your [SCM Private Key] file or paste its contents below."
-            isRequired
-            renderInput={({ field: { onChange, value, name } }) => (
-              <FileUpload
-                data-testid="source-key-upload"
-                id="key"
-                name={name}
-                type="text"
-                value={isKeyEncrypted ? "[Encrypted]" : (value ?? "")}
-                filename={values.keyFilename}
-                filenamePlaceholder="Drag and drop a file or upload one"
-                dropzoneProps={{
-                  onDropRejected: () => setIsKeyFileRejected(true),
-                }}
-                validated={isKeyFileRejected ? "error" : "default"}
-                onFileInputChange={(_, file) => {
-                  setValue("keyFilename", file.name);
-                  setIsKeyFileRejected(false);
-                }}
-                onDataChange={(_, value: string) => {
-                  onChange(value);
-                }}
-                onTextChange={(_, value: string) => {
-                  onChange(value);
-                }}
-                onClearClick={() => {
-                  onChange("");
-                  setValue("keyFilename", "");
-                  setIsKeyFileRejected(false);
-                }}
-                allowEditingUploadedText
-                browseButtonText="Upload"
-              />
-            )}
-          />
-          <HookFormPFTextInput
-            control={control}
-            name="password"
-            fieldId="password"
-            label="Private Key Passphrase"
-            type={isPasswordHidden ? "password" : "text"}
-            formGroupProps={{
-              labelIcon: !isPasswordEncrypted ? (
-                <KeyDisplayToggle
-                  keyName="password"
-                  isKeyHidden={isPasswordHidden}
-                  onClick={toggleHidePassword}
-                />
-              ) : undefined,
-            }}
-            onFocus={() => resetField("password")}
-          />
-        </>
-      )}
-    </>
-  );
-};
-
-const KindAssetForm: React.FC<{
-  identity?: Identity;
-}> = ({ identity }) => {
-  const { control, getValues, setValue, resetField } = useFormContext();
-  const values = getValues();
-
-  const [isKeyFileRejected, setIsKeyFileRejected] = useState(false);
-
-  const [isPasswordHidden, setIsPasswordHidden] = useState(true);
-  const toggleHidePassword = (e: React.FormEvent<HTMLElement>) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsPasswordHidden(!isPasswordHidden);
-  };
-
-  const isPasswordEncrypted = identity?.password === values.password;
-  const isKeyEncrypted = identity?.key === values.key;
-
-  return (
-    <>
-      <HookFormPFGroupController
-        control={control}
-        name="userCredentials"
-        label="User credentials"
-        isRequired
-        fieldId="user-credentials-select"
-        renderInput={({ field: { value, name, onChange } }) => (
-          <SimpleSelect
-            id="user-credentials-select"
-            toggleId="user-credentials-select-toggle"
-            toggleAriaLabel="User credentials select dropdown toggle"
-            aria-label={name}
-            value={
-              value ? toOptionLike(value, USER_CREDENTIALS_OPTIONS) : undefined
-            }
-            options={USER_CREDENTIALS_OPTIONS}
-            onChange={(selection) => {
-              const selectionValue =
-                selection as OptionWithValue<UserCredentials>;
-              onChange(selectionValue.value);
-              // So we don't retain the values from the wrong type of credential
-              resetField("default");
-              resetField("user");
-              resetField("key");
-              resetField("password");
-            }}
-          />
-        )}
-      />
-
-      {values?.userCredentials === "userpass" && (
-        <KindSimpleUsernamePasswordForm
-          identity={identity}
-          usernameLabel="Username"
-          passwordLabel="Password"
-        />
-      )}
-
-      {values?.userCredentials === "source" && (
-        <>
-          <HookFormPFGroupController
-            control={control}
-            name="key"
-            fieldId="key"
-            label="Upload your [SCM Private Key] file or paste its contents below."
-            isRequired
-            renderInput={({ field: { onChange, value, name } }) => (
-              <FileUpload
-                data-testid="source-key-upload"
-                id="key"
-                name={name}
-                type="text"
-                value={isKeyEncrypted ? "[Encrypted]" : (value ?? "")}
-                filename={values.keyFilename}
-                filenamePlaceholder="Drag and drop a file or upload one"
-                dropzoneProps={{
-                  onDropRejected: () => setIsKeyFileRejected(true),
-                }}
-                validated={isKeyFileRejected ? "error" : "default"}
-                onFileInputChange={(_, file) => {
-                  setValue("keyFilename", file.name);
-                  setIsKeyFileRejected(false);
-                }}
-                onDataChange={(_, value: string) => {
-                  onChange(value);
-                }}
-                onTextChange={(_, value: string) => {
-                  onChange(value);
-                }}
-                onClearClick={() => {
-                  onChange("");
-                  setValue("keyFilename", "");
-                  setIsKeyFileRejected(false);
-                }}
-                allowEditingUploadedText
-                browseButtonText="Upload"
-              />
-            )}
-          />
-          <HookFormPFTextInput
-            control={control}
-            name="password"
-            fieldId="password"
-            label="Private Key Passphrase"
-            type={isPasswordHidden ? "password" : "text"}
-            formGroupProps={{
-              labelIcon: !isPasswordEncrypted ? (
-                <KeyDisplayToggle
-                  keyName="password"
-                  isKeyHidden={isPasswordHidden}
-                  onClick={toggleHidePassword}
-                />
-              ) : undefined,
-            }}
-            onFocus={() => resetField("password")}
-          />
-        </>
-      )}
-    </>
-  );
-};
-
-const KindMavenSettingsFileForm: React.FC<{
-  identity?: Identity;
-  defaultIdentities?: Record<IdentityKind, Identity | undefined>;
-}> = ({ identity, defaultIdentities }) => {
-  const { t } = useTranslation();
-  const { control, getValues, setValue } = useFormContext();
-  const values = getValues();
-
-  const [isLoading, setIsLoading] = useState(false);
-  const [isSettingsFileRejected, setIsSettingsFileRejected] = useState(false);
-
-  const isSettingsEncrypted = identity?.settings === values.settings;
-  const kindDefault = defaultIdentities?.[values.kind as IdentityKind];
-  const isReplacingDefault =
-    values.default &&
-    kindDefault &&
-    (!identity || kindDefault.id !== identity.id);
-
-  return (
-    <>
-      <HookFormPFGroupController
-        control={control}
-        name="default"
-        fieldId="default"
-        label="Default credential?"
-        renderInput={({ field: { onChange, value, name, ref } }) => (
-          <>
-            <Switch
-              id="default"
-              name={name}
-              label={t("credentials.default.sourceSwitchLabel")}
-              isChecked={value}
-              onChange={(_, checked) => onChange(checked)}
-              ref={ref}
-            />
-            {isReplacingDefault && (
-              <Alert
-                isInline
-                className="alert-replacing-default"
-                variant="warning"
-                title={t("credentials.default.changeTitle")}
-              >
-                <Trans
-                  i18nKey="credentials.default.mavenChangeWarning"
-                  values={{
-                    name: kindDefault.name,
-                  }}
-                />
-              </Alert>
-            )}
-          </>
-        )}
-      />
-      <HookFormPFGroupController
-        control={control}
-        name="settings"
-        fieldId="settings"
-        label="Upload your Settings file or paste its contents below."
-        isRequired={values.kind === "maven"}
-        errorsSuppressed={isLoading}
-        renderInput={({ field: { onChange, value, name } }) => (
-          <FileUpload
-            data-testid="maven-settings-upload"
-            id="settings"
-            name={name}
-            type="text"
-            value={isSettingsEncrypted ? "[Encrypted]" : (value ?? "")}
-            filename={values.settingsFilename}
-            filenamePlaceholder="Drag and drop a file or upload one"
-            dropzoneProps={{
-              accept: { "text/xml": [".xml"] },
-              onDropRejected: () => setIsSettingsFileRejected(true),
-            }}
-            validated={isSettingsFileRejected ? "error" : "default"}
-            onFileInputChange={(_, file) => {
-              setValue("settingsFilename", file.name);
-              setIsSettingsFileRejected(false);
-            }}
-            onDataChange={(_, value: string) => {
-              onChange(value);
-            }}
-            onTextChange={(_, value: string) => {
-              onChange(value);
-            }}
-            onClearClick={() => {
-              onChange("");
-              setValue("settingsFilename", "");
-              setIsSettingsFileRejected(false);
-            }}
-            onReadStarted={() => setIsLoading(true)}
-            onReadFinished={() => setIsLoading(false)}
-            isLoading={isLoading}
-            allowEditingUploadedText
-            browseButtonText="Upload"
-          />
-        )}
-      />
-    </>
-  );
-};
-
-const KindSimpleUsernamePasswordForm: React.FC<{
-  identity?: Identity;
-  usernameLabel: string;
-  passwordLabel: string;
-  passwordRequired?: boolean;
-}> = ({
-  identity,
-  usernameLabel = "Username",
-  passwordLabel = "Password",
-  passwordRequired = true,
-}) => {
-  const { control, getValues, resetField } = useFormContext();
-  const values = getValues();
-
-  const [isPasswordHidden, setIsPasswordHidden] = useState(true);
-  const toggleHidePassword = (e: React.FormEvent<HTMLElement>) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsPasswordHidden(!isPasswordHidden);
-  };
-
-  const isPasswordEncrypted = identity?.password === values.password;
-
-  return (
-    <>
-      <HookFormPFTextInput
-        control={control}
-        name="user"
-        label={usernameLabel}
-        fieldId="user"
-        isRequired
-      />
-      <HookFormPFTextInput
-        control={control}
-        name="password"
-        label={passwordLabel}
-        fieldId="password"
-        isRequired={passwordRequired}
-        type={isPasswordHidden ? "password" : "text"}
-        formGroupProps={{
-          labelIcon: !isPasswordEncrypted ? (
-            <KeyDisplayToggle
-              keyName="password"
-              isKeyHidden={isPasswordHidden}
-              onClick={toggleHidePassword}
-            />
-          ) : undefined,
-        }}
-        onFocus={() => resetField("password")}
-      />
-    </>
-  );
-};
-
-const KindBearerTokenForm: React.FC<{ identity?: Identity }> = ({
-  identity,
-}) => {
-  const { control, getValues, resetField } = useFormContext();
-  const values = getValues();
-
-  const [isKeyHidden, setIsKeyHidden] = useState(true);
-  const toggleHideKey = (e: React.FormEvent<HTMLElement>) => {
-    e.preventDefault();
-    e.stopPropagation();
-    setIsKeyHidden(!isKeyHidden);
-  };
-
-  const isKeyEncrypted = identity?.key === values.key;
-
-  return (
-    <HookFormPFTextInput
-      control={control}
-      name="key"
-      label={"Token"}
-      fieldId="key"
-      isRequired={true}
-      type={isKeyHidden ? "password" : "text"}
-      formGroupProps={{
-        labelIcon: !isKeyEncrypted ? (
-          <KeyDisplayToggle
-            keyName="key"
-            isKeyHidden={isKeyHidden}
-            onClick={toggleHideKey}
-          />
-        ) : undefined,
-      }}
-      onFocus={() => resetField("key")}
-    />
   );
 };
